requirements.txt
flask
python-telegram-bot==20.7
python-dotenv
